move TILE_SIZE to separate constants file
