+++ assets/scripts/app/Palette.jsx
@@ -5,7 +5,8 @@ import Scrollable from '../ui/Scrollable'
 import { connect } from 'react-redux'
 import { generateRandSeed } from '../util/random'
 import { getAllSegmentInfo } from '../segments/info'
-import { TILE_SIZE, getVariantInfoDimensions } from '../segments/view'
+import { TILE_SIZE } from '../segments/constants'
+import { getVariantInfoDimensions } from '../segments/view'
 import Segment from '../segments/Segment'
 import UndoRedo from './UndoRedo'
