+++ assets/scripts/segments/Segment.jsx
@@ -4,7 +4,8 @@ import { connect } from 'react-redux'
 import MeasurementText from '../ui/MeasurementText'
 import { getSegmentVariantInfo, getSegmentInfo } from '../segments/info'
 import { normalizeSegmentWidth, RESIZE_TYPE_INITIAL, suppressMouseEnter } from './resizing'
-import { drawSegmentContents, getVariantInfoDimensions, segmentsChanged, TILE_SIZE } from './view'
+import { TILE_SIZE } from './constants'
+import { drawSegmentContents, getVariantInfoDimensions, segmentsChanged } from './view'
 import { SETTINGS_UNITS_METRIC } from '../users/localization'
 import { infoBubble, INFO_BUBBLE_TYPE_SEGMENT } from '../info_bubble/info_bubble'
 import { t } from '../app/locale'
