+++ assets/scripts/segments/constants.js
@@ -1,0 +1,2 @@
+export const TILESET_POINT_PER_PIXEL = 2.0
+export const TILE_SIZE = 12 // pixels
